MFC r302788, r303026, r305471
r302788:
Fix the type used to hold the value returned from getopt. On arm64 char is
unsigned so will never be -1.

r303026:
Add missing flags from acpidump. These are defined in the header, but not
printed. The HW_REDUCED flag is useful as it should be set on arm64 to
comply with the ARM Server Base Boot Requirements.

r305471:
Teach acpidump how to parse ACPI 5.1 tables found on the development
ThunderX units in the netperf cluster.


git-svn-id: https://svn.freebsd.org/base/stable/11@317377 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f