MFC: r316667 Fix the NFSv4 client hndling of a stale write verifier in the Commit operation.
When the NFSv4 client Commit operation encountered a stale write verifier,
it erroneously mapped that to EIO. This could have caused recently written
data to be lost when a server crashes/reboots between an UNSTABLE write
and the subsequent commit. This patch fixes this.
The bug was only for the NFSv4 client and did not affect NFSv3.


git-svn-id: https://svn.freebsd.org/base/stable/11@317418 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f