MFC: r316719 Don't throw away Open state when a NFSv4.1 client recovery fails.
If the ExchangeID/CreateSession operations done by an NFSv4.1 client
after the server crashes/reboots fails, it is possible that some process/thread
is waiting for an open_owner lock. If the client state is free'd, this
can cause a crash.
This would not normally happen, but has been observed on a mount of the
AmazonEFS service.


git-svn-id: https://svn.freebsd.org/base/stable/10@317475 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f