MFC: r316782 Add call to svcpool_close() for the NFSv4 callback pool (svcpool_nfscbd).
A function called svcpool_close() was added to the server side krpc by
r313735, so that a pool could be closed without destroying the data structures.
This little patch adds a call to it for the callback pool (svcpool_nfscbd),
so that the nfscbd daemon can be killed/restarted and continue to work
correctly.


git-svn-id: https://svn.freebsd.org/base/stable/11@317519 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f