MFC: r316792 Add an NFSv4.1 mount option for "use one openowner".
Some NFSv4.1 servers such as AmazonEFS can only support a small fixed number
of open_owner4s. This patch adds a mount option called "oneopenown" that
can be used for NFSv4.1 mounts to make the client do all Opens with the
same open_owner4 string. This option can only be used with NFSv4.1 and
may not work correctly when Delegations are is use.

Differential Revision:	https://reviews.freebsd.org/D8988


git-svn-id: https://svn.freebsd.org/base/stable/10@317525 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f