MFC r316776 (by cem@):
linux_ioctl: Refactor some v4l2 struct converters

According to the C standard, it is invalid to copy beyond the end of an
object, even if that object is obviously a member of a larger object (a
struct, in this case).

Appease the standard and Coverity by refactoring the copy in a
straightforward way.  No functional change.


git-svn-id: https://svn.freebsd.org/base/stable/11@317590 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f