MFC: r317350 Fix the default uid/gid values in nfsuserd.c
This patch sets the default uid/gid values for "nobody" and "nogroup"
to the values in the password and group databases. Normally nfsuserd(8)
will override these with whatever is in the password/group databases,
so these values are only used when the databases entries aren't available.
It would be nice to use the definitions in sys/conf.h, but those are
in the _KERNEL section of the file.


git-svn-id: https://svn.freebsd.org/base/stable/11@317919 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f