MFC: r317272 Add checks for failed operations to the NFSv4 client function nfscl_mtofh().
The nfscl_mtofh() function didn't check for failed operations and, as such,
would have returned EBADRPC for these cases, due to parsing failure.
This patch adds checks, so that it returns with ND_NOMOREDATA set.
This is needed for future use in the pNFS server and acts as a safety
belt in the meantime.


git-svn-id: https://svn.freebsd.org/base/stable/11@317924 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f