MFC: r317305 Fix the NFSv4.1/pNFS client return layout on close.
The "return layout on close" case in the pNFS client was badly broken.
Fortunately, extant pNFS servers that I have tested against do not
do this. This patch fixes it. It also changes the way the layout stateid.seqid
is set for LayoutReturn. I think this change is correct w.r.t. the RFC,
but I am not 100% sure.
This was found during recent testing of the pNFS server under development.


git-svn-id: https://svn.freebsd.org/base/stable/10@317978 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f