MFC: r317578
Fix a bug introduced as part of r287726; use the right device_t for
determining the softc of the bridge in psycho_route_interrupt(). [1]
While at it, update the corresponding comment that the code in
question is also necessary for U30s in addition to E450s (a fact
that has been known for ages).

PR:		218478 [1]
Submitted by:	Yoshihiko Iwama [1]


git-svn-id: https://svn.freebsd.org/base/stable/11@318272 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f