MFC r317660, r317710
r317660:
Support clnt_raw's use of FD_SETSIZE as a fake file descriptor.

Accomplish this by allocating space for it in __svc_xports and allowing
it to be registered.  The failure to allocate space was causing an
out-of-bounds read in svc_getreq_common().  The failure to register
caused PR 211804.

The bug was found with CHERI bounds checking.

PR:		211804
Obtained from:	CheriBSD
Sponsored by:	DARPA, AFRL
Reviewed by:	ngie
Differential Revision:	https://reviews.freebsd.org/D10528

r317710:
Remove expected failure now that it was fixed in r317660.

PR:		211804
Reviewed by:	ngie
Obtained from:	CheriBSD
Sponsored by:	DARPA, AFRL
Differential Revision:	https://reviews.freebsd.org/D10576


git-svn-id: https://svn.freebsd.org/base/stable/11@318322 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f