MFC 318360: Fix p_endcopy.
For p_endcopy to work correctly, it must be the name of the next element
in struct proc after the end of the copy region, not the name of the
last element in the copy region.  Currently, the last element
(p_elf_flags) is not being copied.  In addition, the p_xexit and
p_xsig fields should not be copied on fork, so move them out of the
copy region.

Note that for stable/11 the fix is a bit simpler than in HEAD as it
merely consists of formally moving p_xexit and p_xsig out of the
copy region by shrinking the end of the copy region.


git-svn-id: https://svn.freebsd.org/base/stable/11@318563 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f