MFC rr317934:
Add resizewin -z. It makes resizewin not do anything if the terminal
size is already set to something other than zero. It's supposed to be
called from eg /etc/profile - it's not neccessary to query terminal
size when logging in over the network, because the protocol used already
takes care of this, but it's neccessary when logging over a serial line.


git-svn-id: https://svn.freebsd.org/base/stable/11@318728 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f