move p_sigqueue to the end of struct proc
In order to preserve KBI in stable branches, replace the existing
p_sigqueue slot with padding and move the expanded (as of r315949)
p_sigqueue to the end of the struct.

This is a repeat of r317529 (which concerned td_sigqueue in struct
thread) for p_sigqueue in struct proc.

Virtualbox modules (and possibly others) are affected without this fix.

Reviewed by:	kib
Differential Revision:	https://reviews.freebsd.org/D10843


git-svn-id: https://svn.freebsd.org/base/stable/10@318743 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f